Maintainability... Moved pyqt widgets to backend_ui.py and renamed darkflow.py to backend.py
