+++ resources.qrc
@@ -1,6 +1,6 @@
 <!DOCTYPE RCC><RCC version="1.0">
 <qresource>
-<!-- TODO: add strings and translations for libs/darkflow.py -->
+<!-- TODO: add strings and translations for libs/backend.py -->
 <!-- resource and locale-specific resource paths -->
 <file alias="help">resources/icons/help.png</file>
 <file alias="app">resources/icons/app.png</file>
